requirements.txt
Django==4.1.7
djangorestframework==3.15.1
typesense==1.1.1
